refactor(launch): Deduplication
Removed some duplicate code for widget instantiation and added the 'Loading' text to the base function, such that it is more widely used.
